+++ Changelog.md
@@ -3,6 +3,7 @@
 
 ## 1.1.0
 - Assembly info updated.
+- Clear any pending interrupt at startup.
 
 ## 1.0.0
 - Supports CSpect 2.12.26 ClearAllBreakpoints.